Revert "Always generate a chash Value if it not exists in cache"
This reverts commit 99096a6bf5b9f62e6b06a0df398defda3a7d1512.

This code generates a PHP Warning (generateCHash expects string,
not array) and therefor does not work as expected. I've talked to
Axel and he thought this commit was reverted already because it
could possibly flood the TYPO3 cache. Congstar found a different
solution for the problem.
So it seems no one will miss it.